Style: 불필요한 주석 및 파일 정리
