+++ src/App.js
@@ -10,11 +10,12 @@ import MainPage from "./components/MainPage/MainPage";
 import LoginPage from "./components/LoginPage/LoginPage";
 import RegisterPage from "./components/RegisterPage/RegisterPage";
 import ChatPage from "./components/ChatPage/ChatPage";
-import BoardWrite from "components/MainPage/Body/Board/Write";
+import BoardWrite from "components/MainPage/Body/Board/Write2";
 import firebase from "./firebase";
 import { useDispatch, useSelector } from "react-redux";
 import { setUser } from "./redux/actions/user_action";
-import Page from "components/MainPage/Body/Board/Page";
+import Page from "components/MainPage/Body/Board/Write";
+import Write from "components/MainPage/Body/Board/Write";
 
 function App(props) {
   let navigate = useNavigate();
@@ -43,8 +44,8 @@ function App(props) {
       <Route path="/chat" element={<ChatPage />} />
       <Route path="/login" element={<LoginPage />} />
       <Route path="/register" element={<RegisterPage />} />
-      <Route path="/write" element={<BoardWrite />} />
-      <Route path="/page" element={<Page />} />
+      {/* <Route path="/write" element={<BoardWrite />} /> */}
+      <Route path="/write" element={<Write />} />
     </Routes>
   );
 }
